All-sites permission is required for accessing subordinate frames
Chrome does not allow access to subordinate frames without the all-sites
permission. IMO it makes more sense to simply have it from the start,
rather than needing to ask for it at the point of needing to handle
iframes, as there will be many non-obvious cases where this may occur,
and asking will simply confuse the user.