pass the unmodified login path to the native component instead of assuming .gpg
this is in support of
<https://github.com/browserpass/browserpass-native/issues/127>.

this has immediate benefit for anyone using the patches shared in that PR
today. without this, browserpass doesn't recognize `github.com.age` as a
default key for `https://github.com`, because it fails the substring
match. by stripping the extension -- whatever it is -- both
`github.com.gpg` and `github.com.age` are recognized as keys for their
intended domain.